+++ CHANGELOG.md
@@ -20,6 +20,8 @@ info.
 - Multi-instance auto complete setting not being properly read
 - `install.sh` virtual environment issues
 - Fix pool ingress data calls with remotefs (#62)
+- Move additional node prep commands to last set of commands to execute in
+start task (#63)
 
 ## [2.6.0b2] - 2017-03-22
 ### Added